requirements.txt
pandas
polars
numpy
pyarrow
fastparquet
tiktoken
openai
tqdm
python-dotenv